refactor: organize CLI commands by responsibility
